Removed unnecessary functionality from engine
