+++ stubs/contraction_optimization.py
@@ -1,10 +1,13 @@
 from tnreason import engine
 
+alphabet = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't',
+            'u', 'v', 'w', 'x', 'y', 'z']
 
 class ContractionOptimization:
     """
     To manipulate the contraction order and use the other contractors as atomic providers.
     """
+
     def __init__(self, coreDict={}, openColors=[], variableNestedList=None,
                  coreType="NumpyTensorCore"):
         self.coreDict = {key: coreDict[key].clone() for key in coreDict}
