+++ stubs/slice_contractor.py
