+++ tnreason/engine/subscript_creation.py
@@ -1,3 +1,5 @@
+defaultSymbols = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
+
 def get_colorDict(nestedColorsList, symbols):
     colorDict = {}
     i = 0
@@ -10,7 +12,7 @@ def get_colorDict(nestedColorsList, symbols):
                 i += 1
     return colorDict
 
-def get_substring(conCoreDict, openVariables, symbols = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"):
+def get_einsum_substring(conCoreDict, openVariables, symbols = defaultSymbols):
 
     colorDict = get_colorDict([conCoreDict[key].colors for key in conCoreDict], symbols=symbols)
     coreOrder = list(conCoreDict.keys())
@@ -22,3 +24,8 @@ def get_substring(conCoreDict, openVariables, symbols = "abcdefghijklmnopqrstuvw
     rightString = "".join([colorDict[color] for color in colorOrder if color in openVariables])
 
     return "->".join([leftString, rightString]), coreOrder, colorDict, colorOrder
+
+def get_reorder_substring(oldColors, newColors, symbols = defaultSymbols):
+    colorDict = {oldColors[i]: symbols[i] for i in range(len(oldColors))}
+    return "->".join(["".join([colorDict[color] for color in oldColors]),
+                      "".join([colorDict[color] for color in newColors])])
