+++ tnreason/engine/workload_to_tensorflow.py
@@ -1,7 +1,6 @@
 import tensorflow as tf
 
-from tnreason.engine import numpy_contractor as cor
-
+from tnreason.engine import workload_to_numpy as cor
 from tnreason.engine import subscript_creation as subc
 
 
@@ -16,7 +15,7 @@ class TensorFlowCore:
 
 
 class TensorFlowContractor:
-    def __init__(self, coreDict={}, openColors=[]):
+    def __init__(self, coreDict, openColors):
         self.tensorFlowCores = {
             key: TensorFlowCore(values=coreDict[key].values, colors=coreDict[key].colors, name=coreDict[key].name) for
             key in coreDict
@@ -24,22 +23,7 @@ class TensorFlowContractor:
         self.openColors = openColors
 
     def einsum(self):
-        substring, coreOrder, colorDict, colorOrder = subc.get_substring(self.tensorFlowCores, self.openColors)
+        substring, coreOrder, colorDict, colorOrder = subc.get_einsum_substring(self.tensorFlowCores, self.openColors)
         return TensorFlowCore(values=tf.einsum(substring,
                                                *[self.tensorFlowCores[key].values for key in coreOrder]
                                                ), colors=[color for color in colorOrder if color in self.openColors])
-
-
-if __name__ == "__main__":
-    import tnreason.tensor.formula_tensors as ft
-
-    cores = ft.FormulaTensor(["c", "and", ["not", ["a", "or", "b"]]]).get_cores()
-    print(cores)
-
-    contractor = TensorFlowContractor(cores, openColors=["a", "b"])
-
-    result = contractor.einsum()
-    print(result.values)
-    print(result.colors)
-
-    print(result.to_NumpyTensorCore())
