+++ tnreason/engine/workload_to_torch.py
@@ -1,13 +1,12 @@
 import torch as tor
 
-from tnreason.engine import numpy_contractor as cor
-
+from tnreason.engine import workload_to_numpy as cor
 from tnreason.engine import subscript_creation as subc
 
 
 class TorchCore:
-    def __init__(self, values, colors, name=None, inType="Numpy"):
-        if inType == "Numpy":
+    def __init__(self, values, colors, name=None, inType="NumpyTensorCore"):
+        if inType == "NumpyTensorCore":
             self.values = tor.from_numpy(values)
         else:
             self.values = values
@@ -19,7 +18,7 @@ class TorchCore:
 
 
 class TorchContractor:
-    def __init__(self, coreDict={}, openColors=[]):
+    def __init__(self, coreDict, openColors):
         self.torchCores = {
             key: TorchCore(values=coreDict[key].values, colors=coreDict[key].colors, name=coreDict[key].name) for
             key in coreDict
@@ -27,23 +26,8 @@ class TorchContractor:
         self.openColors = openColors
 
     def einsum(self):
-        substring, coreOrder, colorDict, colorOrder = subc.get_substring(self.torchCores, self.openColors)
+        substring, coreOrder, colorDict, colorOrder = subc.get_einsum_substring(self.torchCores, self.openColors)
         return TorchCore(values=tor.einsum(substring,
                                            *[self.torchCores[key].values for key in coreOrder]
                                            ), colors=[color for color in colorOrder if color in self.openColors],
                          inType="Other")
-
-
-if __name__ == "__main__":
-    import tnreason.tensor.formula_tensors as ft
-
-    cores = ft.FormulaTensor(["c", "and", ["not", ["a", "or", "b"]]]).get_cores()
-    print(cores)
-
-    contractor = TorchContractor(cores, openColors=["a", "b"])
-
-    result = contractor.einsum()
-    print(result.values)
-    print(result.colors)
-
-    print(result.to_NumpyTensorCore())
